+++ app/src/main/java/com/codesoom/assignment/auth/application/AuthenticationService.java
@@ -3,8 +3,8 @@ package com.codesoom.assignment.auth.application;
 import com.codesoom.assignment.auth.application.exception.LoginFailException;
 import com.codesoom.assignment.auth.application.port.in.AuthenticationUseCase;
 import com.codesoom.assignment.common.utils.JwtUtil;
-import com.codesoom.assignment.role.repository.Role;
-import com.codesoom.assignment.role.repository.RoleRepository;
+import com.codesoom.assignment.role.application.port.out.RoleRepository;
+import com.codesoom.assignment.role.domain.Role;
 import com.codesoom.assignment.user.application.port.out.UserRepository;
 import com.codesoom.assignment.user.domain.User;
 import io.jsonwebtoken.Claims;
