+++ app/src/main/java/com/codesoom/assignment/common/filters/JwtAuthenticationFilter.java
@@ -2,7 +2,7 @@ package com.codesoom.assignment.common.filters;
 
 import com.codesoom.assignment.auth.application.port.in.AuthenticationUseCase;
 import com.codesoom.assignment.common.security.UserAuthentication;
-import com.codesoom.assignment.role.repository.Role;
+import com.codesoom.assignment.role.domain.Role;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
