--- app/src/main/java/com/codesoom/assignment/role/repository/RoleRepository.java
@@ -1,9 +1,0 @@
-package com.codesoom.assignment.role.repository;
-
-import java.util.List;
-
-public interface RoleRepository {
-    List<Role> findAllByUserId(Long userId);
-
-    Role save(Role role);
-}
