+++ app/src/test/java/com/codesoom/assignment/domain/AuthenticationServiceTest.java
@@ -5,8 +5,8 @@ import com.codesoom.assignment.auth.application.exception.InvalidTokenException;
 import com.codesoom.assignment.auth.application.exception.LoginFailException;
 import com.codesoom.assignment.auth.application.port.in.AuthenticationUseCase;
 import com.codesoom.assignment.common.utils.JwtUtil;
-import com.codesoom.assignment.role.repository.Role;
-import com.codesoom.assignment.role.repository.RoleRepository;
+import com.codesoom.assignment.role.application.port.out.RoleRepository;
+import com.codesoom.assignment.role.domain.Role;
 import com.codesoom.assignment.user.application.port.out.UserRepository;
 import com.codesoom.assignment.user.domain.User;
 import org.junit.jupiter.api.BeforeEach;
